Move check_time_tensor to _utils and use in pwc_pulse
