+++ dynamiqs/_utils.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from typing import Any
 
 import torch
+from torch import Tensor
 
 
 def type_str(type: Any) -> str:
@@ -28,3 +29,20 @@ def to_device(device: str | torch.device | None) -> torch.device:
             'Argument `device` must be a string, a `torch.device` or `None` but has'
             f' type {obj_type_str(device)}.'
         )
+
+
+def check_time_tensor(x: Tensor, arg_name: str, allow_empty=False):
+    # check that a time tensor is valid (it must be a 1D tensor sorted in strictly
+    # ascending order and containing only positive values)
+    if x.ndim != 1:
+        raise ValueError(
+            f'Argument `{arg_name}` must be a 1D tensor, but is a {x.ndim}D tensor.'
+        )
+    if not allow_empty and len(x) == 0:
+        raise ValueError(f'Argument `{arg_name}` must contain at least one element.')
+    if not torch.all(torch.diff(x) > 0):
+        raise ValueError(
+            f'Argument `{arg_name}` must be sorted in strictly ascending order.'
+        )
+    if not torch.all(x >= 0):
+        raise ValueError(f'Argument `{arg_name}` must contain positive values only.')
